+++ library.rb
@@ -1,3 +1,5 @@
+require_relative 'create_functions'
+
 class Library
   include CreateFunctions
   attr_accessor :book_list, :people, :rentals
@@ -30,35 +32,6 @@ class Library
       puts "[#{person.class}] Name: #{person.name}. ID: #{person.id}, Age: #{person.age}"
     end
     puts ''
-  end
-
-  def create_rental
-    check_empty(@book_list, 'Books', '4')
-
-    puts 'Select a book from the following list by number:'
-    @book_list.map do |book|
-      puts "#{@book_list.index(book)}) Title: '#{book.title}', Author: '#{book.author}'"
-    end
-
-    print 'Book #: '
-    book_selection = gets.chomp.to_i
-    book = @book_list[book_selection]
-
-    puts 'Select a person from the following list by number (not ID):'
-    @people.map do |person|
-      puts "#{@people.index(person)}) [#{person.class}] Name: #{person.name}. ID: #{person.id}, Age: #{person.age}"
-    end
-
-    print 'Person #: '
-    person_selection = gets.chomp.to_i
-    person = @people[person_selection]
-
-    print 'Date: '
-    date = gets.chomp.to_i
-
-    rental = Rental.new(date, book, person)
-    @rentals.push(rental)
-    puts 'Rental created successfully'
   end
 
   def list_rentals
